Migrate actions

+++ app/TransactionRules/Actions/ClearBudget.php
@@ -23,6 +23,9 @@ declare(strict_types=1);
 namespace FireflyIII\TransactionRules\Actions;
 
 use DB;
+use FireflyIII\Events\TriggeredAuditLog;
+use FireflyIII\Models\RuleAction;
+use FireflyIII\Models\TransactionJournal;
 use Log;
 
 /**
@@ -30,14 +33,37 @@ use Log;
  */
 class ClearBudget implements ActionInterface
 {
+
+    private RuleAction $action;
+
+    /**
+     * TriggerInterface constructor.
+     *
+     * @param RuleAction $action
+     */
+    public function __construct(RuleAction $action)
+    {
+        $this->action = $action;
+    }
+
     /**
      * @inheritDoc
      */
     public function actOnArray(array $journal): bool
     {
+        /** @var TransactionJournal $journal */
+        $journal = TransactionJournal::where('user_id', $journal['user_id'])->find($journal['transaction_journal_id']);
+        $budget  = $journal->budgets()->first();
+        if (null === $budget) {
+            Log::debug(sprintf('RuleAction ClearBudget, no budget in journal #%d.', $journal['transaction_journal_id']));
+            return false;
+        }
+
         DB::table('budget_transaction_journal')->where('transaction_journal_id', '=', $journal['transaction_journal_id'])->delete();
 
-        Log::debug(sprintf('RuleAction ClearBudget removed all budgets from journal %d.', $journal['transaction_journal_id']));
+        event(new TriggeredAuditLog($this->action->rule, $journal, 'remove_budget', $budget->name, null));
+
+        Log::debug(sprintf('RuleAction ClearBudget removed all budgets from journal #%d.', $journal['transaction_journal_id']));
 
         return true;
     }
